requirements-simplified.txt
# Simplified Requirements for MB-Sparrow (10-user deployment)
# This file contains minimal dependencies without Redis/Celery overhead

# Core Application & Agent Dependencies
fastapi==0.115.5
slowapi>=0.1.9  # Simple in-memory rate limiting
uvicorn==0.29.0
python-multipart>=0.0.6
pydantic==2.11.5
pydantic-settings
email-validator>=2.0.0
python-dotenv

# Langchain Stack
langchain==0.3.25
langchain-core==0.3.68
langchain-community==0.3.25
langchain-google-genai==2.0.8
langgraph

# Google AI SDK
google-generativeai==0.8.3

# Torch - Removed for simplified deployment
# torch==2.0.1  # Not needed for 10 users
# torchvision==0.15.2  # Not needed for 10 users

# Database & Auth
supabase>=2.0.0
gotrue>=2.0.0
qdrant-client
PyJWT>=2.8.0
python-jose[cryptography]==3.3.0
cryptography>=41.0.0
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
pgvector

# External Tools
tavily-python==0.3.0
firecrawl-py

# HTML/PDF Processing
beautifulsoup4==4.12.3
lxml==5.3.0
bleach==6.1.0
pdfplumber==0.11.4
pypdf==5.1.0
reportlab==4.2.5
# pytesseract==0.3.10  # OCR not needed for 10 users
Pillow>=10.3.0  # Keep for basic image handling
# easyocr==1.7.2  # Heavy OCR library not needed

# Agent Dependencies (Simplified)
# faiss-cpu  # Vector DB not needed for 10 users
# sentence-transformers  # Heavy ML model not needed
typing-extensions

# LLM Dependencies
httpx>=0.25.0
aiohttp
openai

# Observability & Logging (simplified)
structlog
loguru

# Utilities
tiktoken  # Keep for token counting
# transformers  # Heavy ML library not needed
# chromadb  # Vector DB not needed for 10 users
pypika==0.48.9
scipy
numpy
pandas
requests
websockets
itsdangerous
passlib
pyyaml
chardet
sqlparse
psutil
python-dateutil
click
rich
faker
nltk>=3.8
networkx>=3.0
scikit-learn>=1.3.0
tenacity>=8.2.0
pdf2image>=1.16.0

# Testing
pytest
pytest-asyncio==0.23.7
pytest-cov==5.0.0
pytest-benchmark==4.0.0